requirements.txt
# MarketMate Backend Requirements

# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
SQLAlchemy==2.0.25
pymysql==1.1.0
cryptography==42.0.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Environment Variables
python-dotenv==1.0.0

# CORS
# Already included in fastapi

# Development
pytest==7.4.4
httpx==0.26.0
